backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
networkx==3.1
gitpython==3.1.32
google-generativeai==0.3.0
python-dotenv==1.0.0
httpx==0.24.1
sqlitedict==2.1.0
python-multipart==0.0.6 